requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.10.0=py37_0
argon2-cffi=20.1.0=pypi_0
astor=0.8.1=py37_0
async-generator=1.10=pypi_0
attrs=20.2.0=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
bleach=3.2.1=pypi_0
blinker=1.4=py37_0
brotlipy=0.7.0=py37he774522_1000
ca-certificates=2020.10.14=0
cachetools=4.1.1=py_0
cairo=1.16.0=hc79e3cf_1005
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.14.2=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
clangdev=5.0.0=flang_3
click=7.1.2=py_0
colorama=0.4.3=pypi_0
console_shortcut=0.1.1=4
cryptography=3.1=py37h7a1dbc1_0
cudatoolkit=10.1.243=h74a9793_0
cudnn=7.6.5=cuda10.1_0
cycler=0.10.0=py_2
decorator=4.4.2=py_0
defusedxml=0.6.0=pypi_0
entrypoints=0.3=pypi_0
freetype=2.10.2=hd328e21_0
gast=0.2.2=py37_0
gettext=0.19.8.1=hb01d8f6_1002
giotto-tda-nightly=20200917.19=pypi_0
glib=2.66.0=h31c07e1_0
google-auth=1.21.2=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
grpcio=1.31.0=py37he7da953_0
h5py=2.8.0=py37hf7173ca_0
hdf5=1.8.18=vc14h7a021fe_0
hopcroftkarp=1.2.4=py_0
icc_rt=2019.0.0=h0cc432a_1
icu=67.1=h33f27b4_0
idna=2.10=py_0
importlib-metadata=1.7.0=py37_0
intel-openmp=2020.2=254
ipykernel=5.3.4=pypi_0
ipython=7.18.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
jedi=0.17.2=pypi_0
jinja2=2.11.2=pypi_0
joblib=0.16.0=py_0
jpeg=9d=he774522_0
jsonschema=3.2.0=pypi_0
jupyter-client=6.1.7=pypi_0
jupyter-core=4.6.3=pypi_0
jupyterlab-pygments=0.1.1=pypi_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_1
kiwisolver=1.2.0=py37heaa310e_0
libblas=3.8.0=14_mkl
libcblas=3.8.0=14_mkl
libclang=10.0.1=default_hf44288c_1
libffi=3.2.1=ha925a31_1007
libflang=5.0.0=h6538335_20180525
libiconv=1.16=he774522_0
libpng=1.6.37=ha81a0f5_2
libprotobuf=3.13.0=h200bbdf_0
libtiff=4.1.0=h885aae3_6
lz4-c=1.9.2=h62dcd97_2
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.2.2=py37_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py37h35e8a6e_0
mistune=0.8.4=pypi_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.1.0=py37h45dec08_0
mkl_random=1.1.0=py37h675688f_0
msys2-conda-epoch=20160418=1
nbclient=0.5.0=pypi_0
nbconvert=6.0.3=pypi_0
nbformat=5.0.7=pypi_0
nest-asyncio=1.4.0=pypi_0
networkx=2.5=py_0
notebook=6.1.4=pypi_0
numpy=1.19.1=py37h5510c5b_0
numpy-base=1.19.1=py37ha3acd2a_0
oauthlib=3.1.0=py_0
olefile=0.46=py_0
openmp=5.0.0=vc14_0
openssl=1.1.1h=he774522_0
opt_einsum=3.1.0=py_0
packaging=20.4=pypi_0
pandas=1.1.3=py37ha925a31_0
pandocfilters=1.4.2=pypi_0
parso=0.7.1=pypi_0
pcre=8.44=h6538335_0
persim=0.1.2=1
pickleshare=0.7.5=pypi_0
pillow=7.2.0=py37hc826c6e_1
pip=20.2.2=py37_0
pixman=0.38.0=hfa6e2cd_1003
plotly=4.10.0=pypi_0
prometheus-client=0.8.0=pypi_0
prompt-toolkit=3.0.7=pypi_0
protobuf=3.13.0=py37h6538335_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycairo=1.19.1=py37hd14c632_3
pycparser=2.20=py_2
pyflagser=0.4.1=pypi_0
pygments=2.7.1=pypi_0
pyjwt=1.7.1=py37_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py37h6538335_1
pyqt5-sip=4.19.18=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyrsistent=0.17.3=pypi_0
pysocks=1.7.1=py37_1
python=3.7.9=h60c2a47_0
python-dateutil=2.8.1=py_0
python-igraph=0.8.2=py37hc69c63d_0
python_abi=3.7=1_cp37m
pytz=2020.1=py_0
pywin32=228=pypi_0
pywinpty=0.5.7=pypi_0
pyzmq=19.0.2=pypi_0
qt=5.12.6=hb2cf2c5_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=py_0
retrying=1.3.3=pypi_0
ripser=0.5.3=py37h1fb7aa8_0
rsa=4.6=py_0
scikit-learn=0.23.2=py37hdc70db3_0
scipy=1.5.2=py37h9439919_0
send2trash=1.5.0=pypi_0
setuptools=49.6.0=py37_0
six=1.15.0=py_0
sqlite=3.33.0=h2a8f88b_0
tensorboard=2.2.1=pyh532a8cf_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.1.0=gpu_py37h7db9008_0
tensorflow-base=2.1.0=gpu_py37h55f5790_0
tensorflow-estimator=2.1.0=pyhd54b08b_0
tensorflow-gpu=2.1.0=h0d30ee6_0
termcolor=1.1.0=py37_1
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
texttable=1.6.3=pyh9f0ad1d_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
tornado=6.0.4=py37hfa6e2cd_0
tqdm=4.51.0=pypi_0
traitlets=5.0.4=pypi_0
urllib3=1.25.10=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=0.16.1=py_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=pypi_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
wrapt=1.12.1=py37he774522_1
xz=5.2.5=h62dcd97_1
zipp=3.1.0=py_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h1f3a1b7_2
